Add force option to artifactory delete aql command

+++ src/providers/artifactory/delete/ArtifactoryDeleteAqlCommand.ts
@@ -2,18 +2,19 @@ import { readFileSync } from 'graceful-fs';
 import isValidPath from 'is-valid-path';
 
 import Command from '../../../core/Command';
-import { GlobalOptions, globalOptions } from '../../../core/commandOptions';
+import { GlobalOptions, globalOptions, ForceOption, forceOption } from '../../../core/commandOptions';
 import { runQuery, deleteArtifact } from '../artifactory-utils';
 
 export type ArtifactoryDeleteAqlCommandOptions =
   GlobalOptions
+  & ForceOption
   & {
     // server: string
     // repo: string
     // packageType: string
     apiKey?: string
     // byChecksum?: boolean
-    // force?: boolean
+    force?: boolean
     aql: string
   }
 
@@ -29,13 +30,14 @@ export default class ArtifactoryDeleteAqlCommand implements Command {
           flags: '--api-key <apiKey>',
           description: 'your API Key, as specified on your user profile page in Artifactory',
         },
+        forceOption,
         ...globalOptions,
       ],
     };
   }
 
   async execute(options: ArtifactoryDeleteAqlCommandOptions) {
-    const { aql, logger } = options;
+    const { aql, force, logger } = options;
 
     const query =
       isValidPath(aql)
@@ -46,10 +48,19 @@ export default class ArtifactoryDeleteAqlCommand implements Command {
     const { results } = await runQuery(query, options);
     logger.info(results);
 
+    if (force) {
+      logger.info('Force enabled: packages will really be deleted!'.red);
+    }
+    else {
+      logger.info('Force disabled: packages will NOT be deleted'.yellow)
+    }
+
     for (const item of results) {
       logger.info('Deleting', item);
-      await deleteArtifact(item, options);
-      logger.info('Deleted'.green, item);
+      if (force) {
+        await deleteArtifact(item, options);
+        logger.info('Deleted'.green, item);
+      }
     }
   }
 }
